reactor: separating image editor from image filtering functionalities
