+++ src/components/Filter/index.tsx
@@ -5,9 +5,10 @@ type FilterProps = {
   fieldName: string;
   applyFilter: (e: React.ChangeEvent<HTMLInputElement>, filterName: string) => void;
   value: number;
+  max?: number;
 }
 
-const Filter: React.FC<FilterProps> = ({ label, applyFilter, fieldName, value }) => {
+const Filter: React.FC<FilterProps> = ({ label, applyFilter, fieldName, value, max }) => {
   return (
     <>
       <label htmlFor={fieldName}>{label}</label>
@@ -16,8 +17,7 @@ const Filter: React.FC<FilterProps> = ({ label, applyFilter, fieldName, value })
         name={fieldName}
         id={fieldName}
         min={0}
-        max={100}
-        defaultValue={0}
+        max={max ?? 10}
         value={value}
         onChange={(e) => applyFilter(e, fieldName)}
       />
