+++ src/types/filterOption.type.ts
@@ -1,0 +1,8 @@
+import { fabric } from "fabric";
+
+export type FilterOption = {
+  fieldName: string;
+  displayName: string;
+  filter: fabric.IBaseFilter[];
+  max?: number;
+}
