+++ src/types/filterState.type.ts
@@ -1,0 +1,3 @@
+export type FiltersStateType = {
+  [key: string]: number;
+}
